pyproject.toml
[tool.poetry]
name = "train-tracker"
version = "0.1.0"
description = ""
authors = ["Grayholm <rodiongrayholm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.120.3"
sqlalchemy = "^2.0.44"
alembic = "^1.17.1"
pydantic = {extras = ["email"], version = "^2.12.3"}
pydantic-settings = "^2.11.0"
psycopg2 = "^2.9.11"
asyncpg = "^0.30.0"
black = "^25.9.0"
uvicorn = "^0.38.0"
argon2-cffi = "^25.1.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
ruff = "^0.14.3"
pyright = "^1.1.407"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100